rabbit/listener/requirements.txt
click==7.1.2
Flask==1.1.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
pika==1.2.0
python-dotenv==0.20.0
Werkzeug==1.0.1
